api_framework_mashang_done/requirements.txt
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-base-url
pytest
pyyaml
rsa
